+++ src/protocol/Edelstein.Protocol.Util.Spatial/IRectangle2D.cs
@@ -5,10 +5,10 @@ public interface IRectangle2D : IObject2D
     IPoint2D P1 { get; }
     IPoint2D P2 { get; }
     
-    float Left { get; }
-    float Right { get; }
-    float Top { get; }
-    float Bottom { get; }
+    int Left { get; }
+    int Right { get; }
+    int Top { get; }
+    int Bottom { get; }
 
     float Height { get; }
     float Width { get; }
